+++ unstructured/partition/pdf_image/pdfminer_utils.py
@@ -1,15 +1,10 @@
 import os
-import tempfile
 from typing import BinaryIO, List, Optional, Tuple
 
-from pdfminer.converter import PDFPageAggregator
-from pdfminer.layout import LAParams, LTContainer, LTImage, LTItem, LTTextLine
-from pdfminer.pdfinterp import PDFPageInterpreter, PDFResourceManager
-from pdfminer.pdfpage import PDFPage
-from pdfminer.psexceptions import PSSyntaxError
+import playa
+from paves.miner import LAParams, LTContainer, LTImage, LTItem, LTTextLine, extract_page
 from pydantic import BaseModel
 
-from unstructured.logger import logger
 from unstructured.utils import requires_dependencies
 
 
@@ -18,18 +13,6 @@ class PDFMinerConfig(BaseModel):
     word_margin: Optional[float] = None
     line_margin: Optional[float] = None
     char_margin: Optional[float] = None
-
-
-def init_pdfminer(pdfminer_config: Optional[PDFMinerConfig] = None):
-    rsrcmgr = PDFResourceManager()
-
-    laparams_kwargs = pdfminer_config.model_dump(exclude_none=True) if pdfminer_config else {}
-    laparams = LAParams(**laparams_kwargs)
-
-    device = PDFPageAggregator(rsrcmgr, laparams=laparams)
-    interpreter = PDFPageInterpreter(rsrcmgr, device)
-
-    return device, interpreter
 
 
 def extract_image_objects(parent_object: LTItem) -> List[LTImage]:
@@ -81,47 +64,29 @@ def rect_to_bbox(
     return (x1, y1, x2, y2)
 
 
-@requires_dependencies(["pikepdf", "pypdf"])
+@requires_dependencies("paves")
 def open_pdfminer_pages_generator(
-    fp: BinaryIO, password: Optional[str] = None, pdfminer_config: Optional[PDFMinerConfig] = None
+    fp: Optional[BinaryIO] = None,
+    filename: Optional[str] = None,
+    password: Optional[str] = None,
+    pdfminer_config: Optional[PDFMinerConfig] = None,
 ):
     """Open PDF pages using PDFMiner, handling and repairing invalid dictionary constructs."""
+    laparams_kwargs = pdfminer_config.model_dump(exclude_none=True) if pdfminer_config else {}
+    laparams = LAParams(**laparams_kwargs)
+    if password is None:
+        password = ""  # playa's default
 
-    import pikepdf
+    if fp is None:
+        from functools import partial
 
-    from unstructured.partition.pdf_image.pypdf_utils import get_page_data
-
-    device, interpreter = init_pdfminer(pdfminer_config=pdfminer_config)
-    with tempfile.TemporaryDirectory() as tmp_dir_path:
-        tmp_file_path = os.path.join(tmp_dir_path, "tmp_file")
-        try:
-            pages = PDFPage.get_pages(fp, password=password or "")
-            # Detect invalid dictionary construct for entire PDF
-            for i, page in enumerate(pages):
-                try:
-                    # Detect invalid dictionary construct for one page
-                    interpreter.process_page(page)
-                    page_layout = device.get_result()
-                except PSSyntaxError:
-                    logger.info("Detected invalid dictionary construct for PDFminer")
-                    logger.info(f"Repairing the PDF page {i + 1} ...")
-                    # find the error page from binary data fp
-                    error_page_data = get_page_data(fp, page_number=i)
-                    # repair the error page with pikepdf
-                    with pikepdf.Pdf.open(error_page_data) as pdf:
-                        pdf.save(tmp_file_path)
-                    page = next(PDFPage.get_pages(open(tmp_file_path, "rb")))  # noqa: SIM115
-                    interpreter.process_page(page)
-                    page_layout = device.get_result()
-                yield page, page_layout
-        except PSSyntaxError:
-            logger.info("Detected invalid dictionary construct for PDFminer")
-            logger.info("Repairing the PDF document ...")
-            # repair the entire doc with pikepdf
-            with pikepdf.Pdf.open(fp) as pdf:
-                pdf.save(tmp_file_path)
-            pages = PDFPage.get_pages(open(tmp_file_path, "rb"))  # noqa: SIM115
-            for page in pages:
-                interpreter.process_page(page)
-                page_layout = device.get_result()
-                yield page, page_layout
+        assert filename
+        with playa.open(
+            filename, space="page", password=password, max_workers=min(1, os.cpu_count() // 2)
+        ) as doc:
+            yield from zip(doc.pages, doc.pages.map(partial(extract_page, laparams=laparams)))
+    else:
+        doc = playa.Document(fp, space="page", password=password)
+        for page in doc.pages:
+            page_layout = extract_page(page, laparams)
+            yield page, page_layout